+++ src/calibration/table_feature_export.cpp
@@ -2,22 +2,22 @@
 #include "fstream"
 
 #include <storage/base_encoded_segment.hpp>
-#include "table_export.hpp"
+#include "table_feature_export.hpp"
 #include "storage/vector_compression/compressed_vector_type.hpp"
 
 namespace opossum {
-    TableExport::TableExport(const std::string &path_to_dir) : _path_to_dir(path_to_dir){
+    TableFeatureExport::TableFeatureExport(const std::string &path_to_dir) : _path_to_dir(path_to_dir){
       //Create dir if not exists
       std::filesystem::create_directories(path_to_dir);
     }
 
-    void TableExport::export_table(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
+    void TableFeatureExport::export_table(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
         _export_table_data(table_wrapper);
         _export_column_data(table_wrapper);
         _export_segment_data(table_wrapper);
     }
 
-   void TableExport::_export_table_data(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
+   void TableFeatureExport::_export_table_data(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
       const auto table_name = table_wrapper->get_name();
       const auto row_count = table_wrapper->get_table()->row_count();
       const auto chunk_size = table_wrapper->get_table()->max_chunk_size();
@@ -29,7 +29,7 @@ namespace opossum {
       _table_csv_writer.write_row();
     }
 
-    void TableExport::_export_column_data(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
+    void TableFeatureExport::_export_column_data(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
 
       auto const table = table_wrapper->get_table();
       int column_count = table->column_count();
@@ -48,7 +48,7 @@ namespace opossum {
       }
     }
 
-    void TableExport::_export_segment_data(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
+    void TableFeatureExport::_export_segment_data(std::shared_ptr<const CalibrationTableWrapper> table_wrapper) {
 
       const auto table = table_wrapper->get_table();
       const auto table_name = table_wrapper->get_name();
